v1.17.0

+++ nerve/swap/src/main/java/network/nerve/swap/storage/impl/SwapPairStorageServiceImpl.java
@@ -29,9 +29,15 @@ import io.nuls.core.model.StringUtils;
 import io.nuls.core.rockdb.service.RocksDBService;
 import network.nerve.swap.constant.SwapDBConstant;
 import network.nerve.swap.model.NerveToken;
+import network.nerve.swap.model.po.StringSetPo;
 import network.nerve.swap.model.po.SwapPairPO;
 import network.nerve.swap.storage.SwapPairStorageService;
 import network.nerve.swap.utils.SwapDBUtil;
+
+import java.util.Collection;
+import java.util.Collections;
+import java.util.HashSet;
+import java.util.Set;
 
 /**
  * @author: PierreLuo
@@ -41,6 +47,7 @@ import network.nerve.swap.utils.SwapDBUtil;
 public class SwapPairStorageServiceImpl implements SwapPairStorageService {
 
     private final String baseArea = SwapDBConstant.DB_NAME_SWAP;
+    private final byte[] PAIR_ALL_KEY = SwapDBUtil.stringToBytes("PAIR-ALL");
     private final String KEY_PREFIX = "PAIR-";
     private final String KEY_PREFIX_LP = "PAIRLP-";
 
@@ -62,7 +69,22 @@ public class SwapPairStorageServiceImpl implements SwapPairStorageService {
         int chainId = AddressTool.getChainIdByAddress(address);
         NerveToken tokenLP = po.getTokenLP();
         RocksDBService.put(baseArea + chainId, SwapDBUtil.stringToBytes(KEY_PREFIX_LP + tokenLP.str()), AddressTool.getAddress(address));
-        return SwapDBUtil.putModel(baseArea + chainId, SwapDBUtil.stringToBytes(KEY_PREFIX + address), po);
+        SwapDBUtil.putModel(baseArea + chainId, SwapDBUtil.stringToBytes(KEY_PREFIX + address), po);
+
+        StringSetPo pairSetPo = SwapDBUtil.getModel(baseArea + chainId, PAIR_ALL_KEY, StringSetPo.class);
+        if (pairSetPo == null) {
+            pairSetPo = new StringSetPo();
+            Set<String> set = new HashSet<>();
+            set.add(address);
+            pairSetPo.setCollection(set);
+            SwapDBUtil.putModel(baseArea + chainId, PAIR_ALL_KEY, pairSetPo);
+        } else {
+            boolean add = pairSetPo.getCollection().add(address);
+            if (add) {
+                SwapDBUtil.putModel(baseArea + chainId, PAIR_ALL_KEY, pairSetPo);
+            }
+        }
+        return true;
     }
 
     @Override
@@ -81,6 +103,16 @@ public class SwapPairStorageServiceImpl implements SwapPairStorageService {
         }
         int chainId = AddressTool.getChainIdByAddress(address);
         return SwapDBUtil.getModel(baseArea + chainId, SwapDBUtil.stringToBytes(KEY_PREFIX + address), SwapPairPO.class);
+    }
+
+    @Override
+    public Collection<String> findAllPairs(int chainId) {
+        StringSetPo addressSetPo = SwapDBUtil.getModel(baseArea + chainId, PAIR_ALL_KEY, StringSetPo.class);
+        if (addressSetPo == null) {
+            return Collections.EMPTY_LIST;
+        }
+        Set<String> addressSet = addressSetPo.getCollection();
+        return addressSet;
     }
 
     @Override
@@ -110,7 +142,20 @@ public class SwapPairStorageServiceImpl implements SwapPairStorageService {
             return false;
         }
         int chainId = AddressTool.getChainIdByAddress(address);
-        return RocksDBService.delete(baseArea + chainId, SwapDBUtil.stringToBytes(KEY_PREFIX + address));
+        SwapPairPO pair = this.getPair(address);
+        if (pair != null) {
+            StringSetPo pairSetPo = SwapDBUtil.getModel(baseArea + chainId, PAIR_ALL_KEY, StringSetPo.class);
+            if(pairSetPo != null) {
+                boolean remove = pairSetPo.getCollection().remove(address);
+                if (remove) {
+                    SwapDBUtil.putModel(baseArea + chainId, PAIR_ALL_KEY, pairSetPo);
+                }
+            }
+            NerveToken tokenLP = pair.getTokenLP();
+            RocksDBService.delete(baseArea + chainId, SwapDBUtil.stringToBytes(KEY_PREFIX_LP + tokenLP.str()));
+            RocksDBService.delete(baseArea + chainId, SwapDBUtil.stringToBytes(KEY_PREFIX + address));
+        }
+        return true;
     }
 
 }
